requirements-dev.txt
graphics.py
pyqt5
python-rtmidi
